S03E0104 testing different methods of encoding...
Currently Topaz AI/AA/200% on the non-fake fields is the winner. It's 480p only, has less noise reduction, bad for text, but this hides the missing field the best of all methods. No other AI models or deinterlacers can produce good results. It just unwatchable when a small object pops in and out of existence as it moves vertically. It's like looking through a cracked glass.